# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Franchise Generation API
  description: |
    REST API for the Franchise Generation platform — a multi-brand franchise ecosystem
    connecting brands, candidates, investors, consultants, and real estate developers.

    ## Authentication
    All protected endpoints require a JWT Bearer token obtained from `POST /auth/login`.
    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    ## Base URL
    Production: `https://api.franchisegeneration.be/api/v1`
    Local dev: `http://localhost:3000/api/v1`

    ## Roles
    | Role | Description |
    |------|-------------|
    | `admin` | Full backoffice access |
    | `investor` | Investment portal access |
    | `candidate` | Candidate onboarding portal |
    | `consultant` | Lead management access |
    | `developer` | Real estate developer access |
    | `brand` | Brand management portal |

  version: 1.0.0
  contact:
    name: Franchise Generation
    url: https://franchisegeneration.be
servers:
  - url: /api/v1
    description: Current version
security:
  - BearerAuth: []
...
paths:
...
  /brands/{id}/opportunities:
    get:
      tags: [Brands]
      summary: List brand opportunities
      description: Returns investment/franchise opportunities belonging to a specific brand.
      parameters:
        - $ref: '#/components/parameters/brandId'
      responses:
        '200':
          description: Array of opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Opportunity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /opportunities:
    get:
      tags: [Opportunities]
      summary: List opportunities
      description: Returns the marketplace of franchise/investment opportunities with optional filters.
      security: []
      parameters:
        - name: brand
          in: query
          description: Filter by brand ID
          schema:
            type: string
            example: atelier
        - name: region
          in: query
          description: Filter by region code
          schema:
            type: string
            example: bxl
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [open, closing, closed, pre-launch]
            example: open
        - name: type
          in: query
          description: Filter by opportunity type
          schema:
            type: string
            enum: [franchise, investment, new-brand]
            example: franchise
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated list of opportunities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList'
              example:
                data:
                  - id: op-gent
                    brand: atelier
                    name: "L'Atelier Gand"
                    city: Gent
                    region: fla
                    format: "Boutique 84 m²"
                    openingDate: "Q3 2026"
                    requiredInvest: 320000
                    candidateContrib: 60000
                    financingNeed: "Tax shelter + prêt investisseur"
                    payback: "4,8 ans"
                    status: "Closing imminent"
                    type: franchise
                total: 6
                page: 1
                limit: 20
        '500':
          $ref: '#/components/responses/InternalError'
  /opportunities/{id}:
    get:
      tags: [Opportunities]
      summary: Get opportunity by ID
      description: Returns full detail for a single opportunity.
      security: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Opportunity detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
        '404':
          $ref: '#/components/responses/NotFound'
  /opportunities/{id}/interest:
    post:
      tags: [Opportunities]
      summary: Express interest in an opportunity
      description: Registers the authenticated user's interest in an opportunity, creating a lead or investment intent record.
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: Optional note from the investor/candidate
                  example: "Je souhaite avoir plus d'informations sur le financement disponible."
                budget:
                  type: integer
                  description: Intended investment amount in EUR
                  example: 25000
      responses:
        '201':
          description: Interest registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Your interest has been registered. A consultant will contact you shortly.
                  leadId:
                    type: string
                    example: lead-042
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Interest already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =========================================================================
  # CANDIDATES
  # =========================================================================
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get notifications
      description: Returns the notification feed for the current user, sorted newest-first.
      parameters:
        - name: unread
          in: query
          description: If true, returns only unread notifications
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Paginated notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedList'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
              example:
                data:
                  - id: n1
                    kind: payment
                    brand: atelier
                    title: "Versement reçu · L'Atelier Châtelain"
                    body: "1 248 € versés (échéance Mai)"
                    read: false
                    createdAt: "2026-05-16T09:00:00Z"
                total: 7
                page: 1
                limit: 50
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/{id}/read:
    put:
      tags: [Notifications]
      summary: Mark notification as read
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications/read-all:
    put:
      tags: [Notifications]
      summary: Mark all notifications as read
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                    description: Number of notifications updated
                    example: 4
        '401':
          $ref: '#/components/responses/Unauthorized'

  # =========================================================================
  # REGIONS
  # =========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier
    brandId:
      name: id
      in: path
      required: true
      schema:
        type: string
        enum: [atelier, couq, cookies, mania]
      description: Brand identifier
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: A valid Bearer token is required.
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: The requested resource was not found.
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: INTERNAL_ERROR
            message: An unexpected error occurred. Please try again.
  schemas:
...
    Opportunity:
      type: object
      description: A franchise or investment opportunity on the marketplace
      properties:
        id:
          type: string
          example: op-gent
        brand:
          type: string
          example: atelier
        name:
          type: string
          example: "L'Atelier Gand"
        city:
          type: string
          example: Gent
        region:
          type: string
          example: fla
        format:
          type: string
          example: "Boutique 84 m²"
        openingDate:
          type: string
          example: "Q3 2026"
        requiredInvest:
          type: integer
          example: 320000
          description: Total investment required in EUR
        candidateContrib:
          type: integer
          example: 60000
          description: Candidate personal contribution in EUR
        financingNeed:
          type: string
          example: "Tax shelter + prêt investisseur"
        payback:
          type: string
          example: "4,8 ans"
        status:
          type: string
          example: "Closing imminent"
        description:
          type: string
        imageUrl:
          type: string
          format: uri
          nullable: true
        type:
          type: string
          enum: [franchise, investment, new-brand]
        target:
          type: integer
          nullable: true
          description: Total fundraising target in EUR
        raised:
          type: integer
          nullable: true
          description: Amount raised to date in EUR
        ticketMin:
          type: integer
          nullable: true
          description: Minimum investment ticket in EUR
        roiTarget:
          type: number
          nullable: true
        closingDays:
          type: integer
          nullable: true
          description: Days remaining until closing
        risk:
          type: string
          nullable: true
          enum: [Faible, Modéré, Élevé]
        createdAt:
          type: string
          format: date-time

    # -------------------------------------------------------------------------
...
    Notification:
      type: object
      description: A user notification
      properties:
        id:
          type: string
          example: n1
        userId:
          type: string
        kind:
          type: string
          enum: [payment, report, doc, opp, system]
          example: payment
        brand:
          type: string
          nullable: true
          example: atelier
        title:
          type: string
          example: "Versement reçu · L'Atelier Châtelain"
        body:
          type: string
          example: "1 248 € versés (échéance Mai)"
        read:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2026-05-16T07:00:00Z"

    # -------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Standard error response envelope
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error description
          example: Request validation failed.
        details:
          type: array
          nullable: true
          description: Field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # -------------------------------------------------------------------------
    PaginatedList:
      type: object
      description: Generic paginated list wrapper
      properties:
        data:
          type: array
          items: {}
          description: Array of result items
        total:
          type: integer
          description: Total number of matching records
          example: 42
        page:
          type: integer
          description: Current page number (1-indexed)
          example: 1
        limit:
          type: integer
          description: Number of items per page
          example: 20
        totalPages:
          type: integer
          description: Total number of pages
          example: 3